refact: nostr types and constants
